refactor(buff): 统一buff持续时间使用tick计算并优化相关逻辑
- 将buff持续时间从float类型改为int类型的tick值
- 添加GetTickFromTime方法用于时间转换
- 优化buff刷新和移除逻辑
- 调整相关技能和战斗管理器的调用方式

+++ Assets/Resources/Scripts/Combat/Buffs/BuffCoolDown.cs
@@ -2,7 +2,7 @@ using System;
 
 public class BuffCoolDown : Buff
 {
-    public BuffCoolDown(int id, int skillId, Chess caster, Chess target, float lastTime)
+    public BuffCoolDown(int id, int skillId, Chess caster, Chess target, int lastTime)
      : base(id, skillId, caster, target, lastTime)
     {
     }
