+++ Assets/Resources/Scripts/Combat/Buffs/BuffDamagedAddRate.cs
@@ -2,7 +2,7 @@ using System;
 
 public class BuffDamagedAddRate : Buff
 {
-    public BuffDamagedAddRate(int id, int skillId, Chess caster, Chess target, float lastTime)
+    public BuffDamagedAddRate(int id, int skillId, Chess caster, Chess target, int lastTime)
      : base(id, skillId, caster, target, lastTime)
     {
     }
