+++ Assets/Resources/Scripts/Combat/Buffs/BuffLock.cs
@@ -1,6 +1,6 @@
 public class BuffLock : Buff
 {
-    public BuffLock(int id, int skillId, Chess caster, Chess target, float lastTime)
+    public BuffLock(int id, int skillId, Chess caster, Chess target, int lastTime)
      : base(id, skillId, caster, target, lastTime)
     {
     }
