+++ Assets/Resources/Scripts/Combat/Buffs/BuffNoAction.cs
@@ -1,6 +1,6 @@
 public class BuffNoAction : Buff
 {
-    public BuffNoAction(int id, int skillId, Chess caster, Chess target, float lastTime)
+    public BuffNoAction(int id, int skillId, Chess caster, Chess target, int lastTime)
      : base(id, skillId, caster, target, lastTime)
     {
     }
